crate user

+++ public/js/ajax.js
@@ -21,6 +21,7 @@ function resultAjax(result, type) {
     var resultArr = JSON.parse(result.responseText);
     if (typeof resultArr.errors === 'undefined') {
         toastSuccess(type);
+        return resultArr.success;
     } else {
         toastError(type);
         return resultArr.errors;
@@ -30,32 +31,69 @@ function resultAjax(result, type) {
 function deleteAjax(url) {
     var confirm = window.confirm('Bạn có muốn xóa dòng dữ liệu này! Sau khi xóa dữ liệu không thể khôi phục lại! Cẩn thận!');
     if (confirm) {
-        var ajax = sendAjax(url, '', 'del', 'delete');
-        if (!ajax) {
-            $('#' + element).DataTable().clear().draw(true);
+        if (confirm) {
+            var ajax = $.ajax({
+                url: url,
+                data: { _method: 'delete' },
+                method: 'delete',
+                typeData: 'json',
+                async: false
+            });
+            var status = (ajax.responseJSON).statusCode;
+            switch (status) {
+                case 200:
+                    $('#' + element).DataTable().clear().draw(true);
+                    toastSuccess('del');
+                    break;
+                case 400:
+                    toastError('del');
+                    break;
+                case 405:
+                    toastMessageDanger('Bạn không có quyền để thực hiện chức năng này!');
+                    break;
+                case 423:
+                    toastMessageDanger('Bạn không thể xóa chính bản thân mình!');
+                    break;
+            }
         }
     }
 }
 
 function deleteManyAjax(url) {
-    var confirm = window.confirm('Bạn có muốn xóa dòng dữ liệu này! Sau khi xóa dữ liệu không thể khôi phục lại! Cẩn thận!');
-    if (confirm) {
-        var arrID = [];
-        $("input:checkbox[name=deleteMany]:checked").each(function () {
-            arrID.push($(this).val());
-        });
-        $.ajax({
-            url: url,
-            type: 'DELETE',
-            dataType: 'json',
-            data: { arrID, _method: 'delete' },
-            success: function(){
-                toastSuccess('del');
-                $('#' + element).DataTable().clear().draw(true);
-            },
-            error: function(){
-                toastError('del');
+    var checked = $("input:checkbox[name=deleteMany]:checked");
+    if (checked.length != 0) {
+        var confirm = window.confirm('Bạn có muốn xóa dòng dữ liệu này! Sau khi xóa dữ liệu không thể khôi phục lại! Cẩn thận!');
+        if (confirm) {
+            var arrID = [];
+            $("input:checkbox[name=deleteMany]:checked").each(function () {
+                arrID.push($(this).val());
+            });
+            var ajax = $.ajax({
+                url: url,
+                data: { arrID, _method: 'delete' },
+                method: 'delete',
+                typeData: 'json',
+                async: false
+            });
+            var status = (ajax.responseJSON).statusCode;
+            switch (status) {
+                case 200:
+                    $('#' + element).DataTable().clear().draw(true);
+                    toastSuccess('del');
+                    break;
+                case 400:
+                    toastError('del');
+                    break;
+                case 405:
+                    toastMessageDanger('Bạn không có quyền để thực hiện chức năng này!');
+                    break;
+                case 423:
+                    toastMessageDanger('Bạn không thể xóa chính bản thân mình!');
+                    break;
             }
-        });
+        }
+    } else {
+        return false;
     }
+
 }
